Move get-customers to explore studio-customers
